Start coding version 1.3 (Vine update)

+++ Core/AudioPlayer/VMPAudioPlayer.m
@@ -3,7 +3,7 @@
 //  OnTheFly
 //
 //  Created by cboy on 10/02/25.
-//  Copyright 2010 sumiisan (aframasda.com). All rights reserved.
+//  Copyright 2010 sumiisan (sumiisan.com). All rights reserved.
 //
 
 #import "VMPAudioPlayer.h"
@@ -213,7 +213,7 @@ static VMHash *processPhaseNames_static_ = nil;
 	//
 	
 	
-	if(status)LLog(@"HardwareCodecPolicy set returned status:%ld",status);
+	if(status)LLog(@"HardwareCodecPolicy set returned status:%d",(int)status);
 #endif
 	
 }
